Implement "Create request" block
Implemented the "Create Request/Offer" block.
Added the "Create Request/Offer" block to the categories page, allowing users to initiate a request or offer based on their role.
The displayed text changes dynamically:
"Tutors for private lessons" for students.
"Students for private lessons" for tutors.
The button label also adjusts based on the user role:
"Create Request" for students.
"Create Offer" for tutors.

The "Create Request/Offer" pop-up is not implemented yet because the corresponding component does not exist.
Introduced the "AppDrawer" component, which will be used to display the pop-up in future tasks.
In the upcoming task, the pop-up will be passed as a child to the "AppDrawer" via props.