Implement Category Card with Hover Effect and Navigation
Changed 'CardWithLink' component for Category Card component that displays an icon, title, and a description showing the number of offers available in the category. The card includes a hover effect. Redirects users to the Subjects page of the selected category when clicked, will be implement in future task in 'Category' page.